Add RPRX layer conversion
